+++ packages/eslint-config-base/test/test-base.js
@@ -2,7 +2,7 @@ import fs from 'fs';
 import path from 'path';
 import test from 'tape';
 
-import index from '../';
+import index from '..';
 
 const files = { ...{ index } }; // object spread is to test parsing
 
@@ -20,8 +20,8 @@ Object.keys(files).forEach((
     t.plan(2);
 
     // scan plugins for react and fail if it is found
-    const hasReactPlugin = Object.prototype.hasOwnProperty.call(config, 'plugins') &&
-      config.plugins.indexOf('react') !== -1;
+    const hasReactPlugin = Object.prototype.hasOwnProperty.call(config, 'plugins')
+      && config.plugins.indexOf('react') !== -1;
     t.notOk(hasReactPlugin, 'there is no react plugin');
 
     // scan rules for react/ and fail if any exist